Fix a couple of narrowing conversions and deactivate the check for now
We do have a couple of narrowing conversions left, but we are pretty
sure, they won't affect anything. Therefore, we decided to deactivate
the check for now. With a later version of clang-tidy we can use NOLINT
blocks to deactivate this check for sections of code.

+++ src/control/SVHFingerManager.cpp
@@ -1542,7 +1542,7 @@ float SVHFingerManager::setForceLimit(const SVHChannel& channel, float force_lim
   current_settings = m_current_settings[channel];
 
   current_settings.wmx = current;
-  current_settings.wmn = -current;
+  current_settings.wmn = -static_cast<float>(current);
 
   if (setCurrentSettings(channel, current_settings))
   {
